feat: change to client comp and add suspense for better UX
