+++ web/package.json
@@ -14,6 +14,7 @@
     "next": "14.2.4",
     "react": "^18",
     "react-dom": "^18",
+    "react-loading-skeleton": "^3.4.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
